Morning changes

+++ Extravelocity/src/reservables/air/Flight.java
@@ -98,9 +98,9 @@ public class Flight {
 		return (this.seats.size() == capacity);
 	}
 	
-	// pre: an int people and an int seatClass
+	// pre: an int people and a double seatClass
 	// post: returns true if there are enough seats to accommodate the amount people in the given seatClass
-	public boolean hasEnoughSeats(int people, int seatClass) {
+	public boolean hasEnoughSeats(int people, double seatClass) {
 		int available = 0;
 		for(Seat seat: this.getSeats()) {
 			if(seat.getAvailable() && seat.getPriority() == seatClass) {
